+++ packages/client/ui/react-ui/src/types/wallet.ts
@@ -7,5 +7,6 @@ export {
     type Transaction,
     EVMWallet,
     SolanaWallet,
+    StellarWallet,
     Wallet,
 } from "@crossmint/wallets-sdk";