+++ packages/wallets/src/signers/non-custodial/ncs-stellar-signer.ts
@@ -1,24 +1,22 @@
-import { VersionedTransaction } from "@solana/web3.js";
 import base58 from "bs58";
 import type { EmailInternalSignerConfig, PhoneInternalSignerConfig } from "../types";
-import { NonCustodialSigner, DEFAULT_EVENT_OPTIONS } from "./ncs-signer";
+import { DEFAULT_EVENT_OPTIONS, NonCustodialSigner } from "./ncs-signer";
 
-export class SolanaNonCustodialSigner extends NonCustodialSigner {
+export class StellarNonCustodialSigner extends NonCustodialSigner {
     constructor(config: EmailInternalSignerConfig | PhoneInternalSignerConfig) {
         super(config);
     }
 
     async signMessage() {
-        return await Promise.reject(new Error("signMessage method not implemented for email signer"));
+        return await Promise.reject(new Error("signMessage method not implemented for stellar signer"));
     }
 
     async signTransaction(transaction: string): Promise<{ signature: string }> {
         await this.handleAuthRequired();
         const jwt = this.getJwtOrThrow();
 
-        const transactionBytes = base58.decode(transaction);
-        const deserializedTransaction = VersionedTransaction.deserialize(transactionBytes);
-        const messageData = deserializedTransaction.message.serialize();
+        const transactionBuffer = Buffer.from(transaction, "base64");
+        const transactionBase58 = base58.encode(transactionBuffer);
 
         const res = await this.config.clientTEEConnection?.sendAction({
             event: "request:sign",
@@ -30,7 +28,7 @@ export class SolanaNonCustodialSigner extends NonCustodialSigner {
                 },
                 data: {
                     keyType: "ed25519",
-                    bytes: base58.encode(messageData),
+                    bytes: transactionBase58,
                     encoding: "base58",
                 },
             },
@@ -44,8 +42,11 @@ export class SolanaNonCustodialSigner extends NonCustodialSigner {
         if (res?.signature == null) {
             throw new Error("Failed to sign transaction");
         }
-        SolanaNonCustodialSigner.verifyPublicKeyFormat(res.publicKey);
-        return { signature: res.signature.bytes };
+        StellarNonCustodialSigner.verifyPublicKeyFormat(res.publicKey);
+
+        const signatureBuffer = base58.decode(res.signature.bytes);
+        const signatureBase64 = Buffer.from(signatureBuffer).toString("base64");
+        return { signature: signatureBase64 };
     }
 
     static verifyPublicKeyFormat(publicKey: { encoding: string; keyType: string; bytes: string } | null) {