+++ apps/auth/node-and-express-server/CHANGELOG.md
@@ -1,5 +1,17 @@
 # crossmint-auth-node
 
+## 1.1.48
+
+### Patch Changes
+
+- @crossmint/server-sdk@1.2.37
+
+## 1.1.47
+
+### Patch Changes
+
+- @crossmint/server-sdk@1.2.36
+
 ## 1.1.46
 
 ### Patch Changes