+++ apps/wallets/quickstart-devkit/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @crossmint/wallets-quickstart-devkit
 
+## 0.1.41
+
+### Patch Changes
+
+- 74a6281: Stellar SDK support
+- Updated dependencies [74a6281]
+  - @crossmint/client-sdk-react-ui@2.2.20
+
+## 0.1.40
+
+### Patch Changes
+
+- @crossmint/client-sdk-react-ui@2.2.19
+
+## 0.1.39
+
+### Patch Changes
+
+- Updated dependencies [5495e50]
+- Updated dependencies [be2c5bf]
+- Updated dependencies [946912e]
+  - @crossmint/client-sdk-react-ui@2.2.18
+
 ## 0.1.38
 
 ### Patch Changes