+++ apps/wallets/smart-wallet/expo/CHANGELOG.md
@@ -1,5 +1,33 @@
 # expo-demo
 
+## 1.1.62
+
+### Patch Changes
+
+- 74a6281: Stellar SDK support
+- Updated dependencies [74a6281]
+  - @crossmint/client-sdk-react-native-ui@0.9.5
+  - @crossmint/wallets-sdk@0.11.2
+
+## 1.1.61
+
+### Patch Changes
+
+- Updated dependencies [2ee5dd6]
+  - @crossmint/wallets-sdk@0.11.1
+  - @crossmint/client-sdk-react-native-ui@0.9.4
+
+## 1.1.60
+
+### Patch Changes
+
+- Updated dependencies [5e829bd]
+- Updated dependencies [bb25455]
+- Updated dependencies [be2c5bf]
+- Updated dependencies [bb25455]
+  - @crossmint/client-sdk-react-native-ui@0.9.3
+  - @crossmint/wallets-sdk@0.11.0
+
 ## 1.1.59
 
 ### Patch Changes