+++ packages/client/wallets/smart-wallet/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @crossmint/client-sdk-smart-wallet
 
+## 0.3.28
+
+### Patch Changes
+
+- 74a6281: Stellar SDK support
+- Updated dependencies [74a6281]
+  - @crossmint/common-sdk-base@0.9.1
+  - @crossmint/client-sdk-base@1.5.16
+
+## 0.3.27
+
+### Patch Changes
+
+- Updated dependencies [bb25455]
+- Updated dependencies [bb25455]
+  - @crossmint/common-sdk-base@0.9.0
+  - @crossmint/client-sdk-base@1.5.15
+
 ## 0.3.26
 
 ### Patch Changes