+++ packages/common/auth/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @crossmint/common-sdk-auth
 
+## 1.0.47
+
+### Patch Changes
+
+- Updated dependencies [74a6281]
+  - @crossmint/common-sdk-base@0.9.1
+  - @crossmint/client-sdk-base@1.5.16
+
+## 1.0.46
+
+### Patch Changes
+
+- Updated dependencies [bb25455]
+- Updated dependencies [bb25455]
+  - @crossmint/common-sdk-base@0.9.0
+  - @crossmint/client-sdk-base@1.5.15
+
 ## 1.0.45
 
 ### Patch Changes