+++ packages/server/CHANGELOG.md
@@ -1,5 +1,22 @@
 # @crossmint/server-sdk
 
+## 1.2.37
+
+### Patch Changes
+
+- Updated dependencies [74a6281]
+  - @crossmint/common-sdk-base@0.9.1
+  - @crossmint/common-sdk-auth@1.0.47
+
+## 1.2.36
+
+### Patch Changes
+
+- Updated dependencies [bb25455]
+- Updated dependencies [bb25455]
+  - @crossmint/common-sdk-base@0.9.0
+  - @crossmint/common-sdk-auth@1.0.46
+
 ## 1.2.35
 
 ### Patch Changes